moved main into separate file and made the algorith available as library
