+++ knapsack_test.go
@@ -1,21 +1,21 @@
-package main
+package knapsack
 
 import (
 	"testing"
 )
 
 var cap1 = 20.
-var testItems1 = []item{
-	{name: "clock", value: 175, weight: 10},
-	{name: "painting", value: 90, weight: 9},
-	{name: "radio", value: 20, weight: 4},
-	{name: "vase", value: 50, weight: 2},
-	{name: "book", value: 10, weight: 1},
-	{name: "computer", value: 200, weight: 20},
+var testItems1 = []Item{
+	{Name: "clock", Value: 175, Weight: 10},
+	{Name: "painting", Value: 90, Weight: 9},
+	{Name: "radio", Value: 20, Weight: 4},
+	{Name: "vase", Value: 50, Weight: 2},
+	{Name: "book", Value: 10, Weight: 1},
+	{Name: "computer", Value: 200, Weight: 20},
 }
 
 var cap2 = 26.
-var testItems2 = []item{
+var testItems2 = []Item{
 	{"item1", 24., 12.},
 	{"item2", 13., 7.},
 	{"item3", 23., 11.},
@@ -26,12 +26,12 @@ var testItems2 = []item{
 func TestGreedyDensity6(t *testing.T) {
 
 	f := func(i, j int) bool {
-		return testItems1[i].value/testItems1[i].weight > testItems1[j].value/testItems1[j].weight
+		return testItems1[i].Value/testItems1[i].Weight > testItems1[j].Value/testItems1[j].Weight
 	}
 
 	v := 0.
-	for _, i := range greedy(testItems1, cap1, f) {
-		v += i.value
+	for _, i := range Greedy(testItems1, cap1, f) {
+		v += i.Value
 	}
 
 	if v != 255.0 {
@@ -42,12 +42,12 @@ func TestGreedyDensity6(t *testing.T) {
 func TestGreedyDensity5(t *testing.T) {
 
 	f := func(i, j int) bool {
-		return testItems2[i].value/testItems2[i].weight > testItems2[j].value/testItems2[j].weight
+		return testItems2[i].Value/testItems2[i].Weight > testItems2[j].Value/testItems2[j].Weight
 	}
 
 	v := 0.
-	for _, i := range greedy(testItems2, cap2, f) {
-		v += i.value
+	for _, i := range Greedy(testItems2, cap2, f) {
+		v += i.Value
 	}
 
 	if v != 47.0 {
@@ -57,21 +57,21 @@ func TestGreedyDensity5(t *testing.T) {
 
 func TestOptimal(t *testing.T) {
 
-	v, _ := bestSolution(testItems1, cap1)
+	v, _ := BestSolution(testItems1, cap1)
 
 	if v != 275.0 {
 		t.Error("Expected 275, got ", v)
 	}
 }
 
-func benchmarkDensity(b *testing.B, set []item, cap float64) {
+func benchmarkDensity(b *testing.B, set []Item, cap float64) {
 
 	f := func(i, j int) bool {
-		return testItems1[i].value/testItems1[i].weight > testItems1[j].value/testItems1[j].weight
+		return testItems1[i].Value/testItems1[i].Weight > testItems1[j].Value/testItems1[j].Weight
 	}
 
 	for n := 0; n < b.N; n++ {
-		greedy(set, cap, f)
+		Greedy(set, cap, f)
 	}
 }
